backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
google-generativeai>=0.7.0
chromadb==0.5.0
python-dotenv==1.0.1
pandas==2.2.2
pytest==8.2.2
httpx==0.27.0
python-multipart==0.0.9
scikit-learn==1.5.0
sentence-transformers==2.7.0
